fix the MemManage fault during initial task start
If an EXC_RETURN value is loaded into the PC when in Thread mode, or from the vector table, or by any other
instruction, the value is treated as an address, not as a special value. The 0xFXXXXXXX address range, that includes all
possible EXC_RETURN values, has Execute Never (XN) permissions, and loading this value causes a
MemManage exception, or an INVSTATE UsageFault exception, or escalation of the exception to a HardFault.